requirements.txt
# Core ML / RL
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.0
torch>=2.2.0
tensorboard>=2.16.0

# Data & Features
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b

# Alpaca Markets
alpaca-py>=0.20.0

# Polymarket
py-clob-client>=0.14.0
web3>=6.15.0

# Database
sqlalchemy>=2.0.0

# Dashboard
streamlit>=1.32.0
plotly>=5.18.0

# Scheduler
APScheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
scikit-learn>=1.4.0
joblib>=1.3.0

# Dev / Training utilities
tqdm>=4.66.0
rich>=13.7.0
